Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
groq
pandas
matplotlib
seaborn
scikit-learn
joblib
pydantic[email]
email-validator
httpx
notebook
ipykernel
